requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.3.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==4.2.23
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
idna==3.10
kombu==5.5.4
Markdown==3.8.2
mysqlclient @ file:///E:/Projects%20Django/Django%20REST/Subscription%20Management/mysqlclient-2.2.7-cp312-cp312-win_amd64.whl#sha256=4b4c0200890837fc64014cc938ef2273252ab544c1b12a6c1d674c23943f3f2e
packaging==25.0
prompt_toolkit==3.0.51
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.1
requests==2.31.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13